refactor: Move CatBreedDTO to its own file and add toModel() extension - Moves `CatBreedDTO` to its own file under a new `../DTOs` folder. - Adds a `toModel()` extension to `CatBreedDTO` to encapsulate the logic for converting the DTO to the `CatBreed` model.
